rename classes, edit readme and fix sonnar issues
